Rename integration test
